perf: consolidate PR data fetching onto single smart deduped path
The repo view page fetched the large pull_requests payload up to three
times per view through three parallel code paths. Make fetchPRDataSmart
(via the request-deduplicating wrapper in supabase-pr-data-smart-deduped)
the single canonical fetcher so concurrent callers share one request.

- Migrate use-cached-repo-data, use-fast-pr-data, use-repo-stats, and
  all five insights modules off fetchPRDataWithFallback
- Remove vestigial useSmartFetch feature-flag branching in
  use-cached-repo-data (smart fetch was already always on)
- Point remaining direct supabase-pr-data-smart importers at the
  deduped wrapper
- Delete dead code: supabase-pr-data.ts (legacy fallback path),
  supabase-pr-data-v2.ts (zero importers), use-repo-data.ts (only
  referenced by its own test)

Part of #1815 (Phase 2).
